More code review changes
Make `InstrumentTrackView::determinePixmap` static. I have also attempted to keep the non-static member function version and to use the `InstrumentTrackView`'s model with a cast. However, this did not work because 'setModel' is executed as the very last step in the constructor, i.e. the model is not set when `determinePixmap` is called. Pulling it to the top is too risky right now. It's a little bit strange anyway that the `InstrumentTrackView` immediately "forgets" that it's working on an `InstrumentTrack` and that casts would be necessary although it receives a "pristine" `InstrumentTrack` in it's constructor.

Rename `s_beatStepButtonOffset` to `BeatStepButtonOffset` and define it as a constexpr.

More corrections of whitespace and indentation.