Optimize map accesses in LadspaManager
Some methods in `LadspaManager` performed several searches through the map by first calling `contains` and then by actually fetching the entry. This is fixed by using `find` on the map. It returns an iterator which can directly provide the result or show that nothing was found. That way we only search through the map once.