Remove use of QApplication::desktop
Remove use of `QApplication::desktop` in ComboBox.cpp. The method was already deprecated in Qt5 and is removed in Qt6.

Generalize the check if the menu can be shown in the screen. The code now also does handles the case where the menu would go out of screen along the X axis.